refactor(credit): rename postgresdrive to postgresadapter to conform with naming
