+++ internal/credit/postgresadapter/balance_snapshot.go
